requirements.txt
asn1crypto==0.24.0
Babel==2.5.3
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.2.2
dataclasses==0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==2.0.1
django-allauth==0.34.0
django-crontab==0.7.1
django-jet==1.0.7
django-js-asset==1.0.0
django-mptt==0.9.0
django-phonenumber-field==2.0.0
django-rest-auth==0.9.2
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.6
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10
livereload==2.5.1
MarkupSafe==1.0
nexmo==2.0.0
oauthlib==2.0.6
phonenumberslite==8.8.10
Pillow==5.0.0
psycopg2==2.7.3.2
pycparser==2.18
pyfcm==1.4.3
Pygments==2.2.0
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
six==1.11.0
sqlparse==0.2.4
tornado==4.5.3
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
